Added the engine drone sound and something to make it change its pitch based on how fast we're going.
Fixed a bug in SoundComponent where the "offset" for the position was being read from the .muffin file and not the .thing file.

Also removed the ability to set the maximum distance of a sound, since it turns out that that DOESN'T mean the sound volume is 0 past that difference - it means that irrklang stops changing the volume past that distance. So if you had a small distance, the sound would still be audible from ridiculously far away

git-svn-id: svn://svn.code.sf.net/p/ponykart/code/trunk@428 ba97e771-f6de-4074-9afc-763755853dd4